+++ crates/oxc_transformer/src/es2022/class_properties/private_field.rs
@@ -11,13 +11,17 @@ use oxc_traverse::{
     ast_operations::get_var_name_from_node, Ancestor, BoundIdentifier, TraverseCtx,
 };
 
-use crate::{common::helper_loader::Helper, TransformCtx};
+use crate::{
+    common::helper_loader::Helper,
+    utils::ast_builder::{create_bind_call, create_call_call, create_member_callee},
+    TransformCtx,
+};
 
 use super::{
     class_details::ResolvedGetSetPrivateProp,
     utils::{
-        create_assignment, create_bind_call, create_call_call, create_member_callee,
-        create_underscore_ident_name, debug_assert_expr_is_not_parenthesis_or_typescript_syntax,
+        create_assignment, create_underscore_ident_name,
+        debug_assert_expr_is_not_parenthesis_or_typescript_syntax,
     },
     ClassProperties, ResolvedPrivateProp,
 };
